runpod/requirements.txt
runpod>=1.6.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
gdown>=4.7.0
timm>=0.9.0
einops>=0.7.0
opencv-python-headless>=4.8.0
requests>=2.31.0

